refactor(objectives): move Source extraction contract into analysis
Why:
- Source-local scientific extraction was split across generic prompt, schema, and response-client modules, obscuring the boundary with deterministic grounding.

What:
- Make the analysis stage own its extraction prompt, schema, validation, bounded repair, and model call while retaining shared provider and JSON mechanics.

Impact:
- Source payloads, three-attempt repair, immediate grounding, route failures, persistence, and public contracts remain unchanged.

Verification:
- All 191 focused extraction, comparison, workflow, and response tests passed, along with all 6 collection-build integration tests.

Refs:
- ISSUE: #330
