+++ backend/application/core/objectives/research_objective_service.py
@@ -19,7 +19,10 @@ from application.core.objectives.analysis.evidence_routing import (
 from application.core.objectives.analysis.paper_experiment import (
     reconstruct_paper_experiments,
 )
-from application.core.objectives.analysis.source_extraction import extract_source_facts
+from application.core.objectives.analysis.source_extraction import (
+    ObjectiveSourceExtractor,
+    extract_source_facts,
+)
 from application.core.objectives.analysis.source_screening import (
     ObjectiveSourceScreener,
     screen_sources,
@@ -111,6 +114,7 @@ class ResearchObjectiveService:
         axis_equivalence_classifier: ResearchAxisEquivalenceClassifier | None = None,
         objective_source_screener: ObjectiveSourceScreener | None = None,
         objective_evidence_router: ObjectiveEvidenceRouter | None = None,
+        objective_source_extractor: ObjectiveSourceExtractor | None = None,
         paper_study_window_extractor: PaperStudyWindowExtractor | None = None,
         paper_signal_reconciler: PaperSignalReconciler | None = None,
         paper_facts_extractor: PaperFactsExtractor | None = None,
@@ -120,6 +124,7 @@ class ResearchObjectiveService:
         self._axis_equivalence_classifier = axis_equivalence_classifier
         self._objective_source_screener = objective_source_screener
         self._objective_evidence_router = objective_evidence_router
+        self._objective_source_extractor = objective_source_extractor
         self._paper_study_window_extractor = paper_study_window_extractor
         self._paper_signal_reconciler = paper_signal_reconciler
         self._paper_facts_extractor = paper_facts_extractor
@@ -273,9 +278,13 @@ class ResearchObjectiveService:
             ],
             progress_callback=progress_callback,
         )
+        if self._objective_source_extractor is None:
+            self._objective_source_extractor = ObjectiveSourceExtractor(
+                objective_inputs["extractor"]
+            )
         source_drafts = extract_source_facts(
             collection_id=collection_id,
-            extractor=objective_inputs["extractor"],
+            source_extractor=self._objective_source_extractor,
             paper_facts_extractor=self._paper_facts_extractor,
             objectives=(objective,),
             objective_paper_frames=paper_frames,
