+++ backend/tests/support/fake_domain_model_extractor.py
@@ -11,6 +11,10 @@ from application.core.objectives.analysis.evidence_routing import (
     StructuredEvidenceSelection,
     StructuredEvidenceSelections,
 )
+from application.core.objectives.analysis.source_extraction import (
+    StructuredEvidenceExtraction,
+    StructuredEvidenceExtractions,
+)
 from application.core.objectives.discovery.axis_equivalence import (
     StructuredAxisCanonicalizationPlan,
 )
@@ -18,10 +22,6 @@ from application.core.objectives.discovery.signal_reconciliation import (
     StructuredPaperSignalReconciliation,
 )
 from application.core.objectives.discovery.study_window import StructuredPaperSkim
-from application.core.objectives.schemas import (
-    StructuredEvidenceExtraction,
-    StructuredEvidenceExtractions,
-)
 from application.core.paper_facts.schemas import (
     MeasurementValuePayload,
     StructuredTableBatchMentions,
@@ -430,7 +430,7 @@ class FakeDomainModelExtractor:
             )
         return StructuredEvidenceSelections(selections=routes)
 
-    def extract_objective_evidence(
+    def extract_source(
         self,
         payload: dict[str, Any],
     ) -> StructuredEvidenceExtractions:
