+++ backend/tests/support/research_objective_service.py
@@ -40,6 +40,7 @@ def build_research_objective_service(
     if objective_extractor is not None:
         kwargs.setdefault("axis_equivalence_classifier", objective_extractor)
         kwargs.setdefault("objective_evidence_router", objective_extractor)
+        kwargs.setdefault("objective_source_extractor", objective_extractor)
         kwargs.setdefault("objective_source_screener", objective_extractor)
         kwargs.setdefault("paper_study_window_extractor", objective_extractor)
         kwargs.setdefault("paper_signal_reconciler", objective_extractor)
